completed spliting the backend appropriately, looks more organised and ready for more....but gotten more files :smile: not so mich an issue. supriced to see i have 43 endpoints :smile:
