+++ api/microservices/institution/settings/departments/get_all.js
@@ -1,0 +1,6 @@
+module.exports = (req,res) => {
+  db.proschool.findOne({id:req.params.school_id}).populate("departments").exec((err, university)=>{
+    if(err) throw err;
+    res.send(university)
+  })
+}
