+++ api/microservices/institution/settings/departments/make_new.js
@@ -1,0 +1,6 @@
+module.exports = (req,res) => {
+  db.department.create({name:req.body.name,proschool:req.params.school_id}).exec((err, department)=>{
+    if(err) throw err;
+    res.send(department)
+  })
+}
