+++ api/microservices/institution/settings/schools/get_all.js
@@ -1,0 +1,6 @@
+module.exports = (req,res) => {
+  db.university.findOne({id:req.params.id}).populate("proschools").exec((err, universities)=>{
+    if(err) throw err;
+    res.send(universities)
+  })
+}
