+++ api/microservices/institution/settings/stages/get_all.js
@@ -1,0 +1,6 @@
+module.exports = (req,res) => {
+  db.level.findOne({id:req.params.id}).populate("stages").exec((err, level)=>{
+    if(err) throw err;
+    res.send(level)
+  })
+}
