+++ api/microservices/institution/settings/study_modes/get_all.js
@@ -1,0 +1,6 @@
+module.exports = (req,res) => {
+  db.university.findOne({id:req.params.uniId}).populate("study_sessions").exec((err, universities)=>{
+    if(err) throw err;
+    res.send(universities)
+  })
+}
