+++ api/microservices/institution/settings/tri_semesters/get_all.js
@@ -1,0 +1,6 @@
+module.exports = (req,res) => {
+  db.university.findOne({id:req.params.uniId}).populate("tri_semesters").populate("active_tri_semester").exec((err, universities)=>{
+    if(err) throw err;
+    res.send(universities)
+  })
+}
