+++ api/microservices/institution/settings/tri_semesters/kill_trisem.js
@@ -1,0 +1,6 @@
+module.exports = (req,res) => {
+  db.tri_semester.destroy({id:req.params.tri_semester_id}).exec((err, tri_semester)=>{
+    if(err) throw err;
+    res.send(tri_semester)
+  })
+}
