+++ api/microservices/institution/settings/units/get_available.js
@@ -1,0 +1,6 @@
+module.exports = (req,res) => {
+  db.department.findOne({id:req.params.id}).populate("units").exec((err, level)=>{
+    if(err) throw err;
+    res.send(level)
+  })
+}
