+++ api/microservices/institution/settings/units/get_offered.js
@@ -1,0 +1,6 @@
+module.exports = (req,res) => {
+  db.stage.findOne({id:req.params.id}).populate("units_i_offer").exec((err, level)=>{
+    if(err) throw err;
+    res.send(level)
+  })
+}
