+++ api/microservices/institution/settings/units/get_offered_by_course.js
@@ -1,0 +1,7 @@
+module.exports = (req,res) => {
+  db.stage.findOne({id:req.params.id})
+    .populate("units_i_offer").exec((err, level)=>{
+    if(err) throw err;
+    res.send(level)
+  })
+}
