--- api/routes/basic/static/public/index.html
@@ -1,9 +1,0 @@
-<<!DOCTYPE html>
-<html>
-<head>
-	<title>head</title>
-</head>
-<body>
-awesomness
-</body>
-</html>
